% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_themeSwitcher.R
\name{ChangeThemeOutputUI}
\alias{ChangeThemeOutputUI}
\title{Theme Switcher - theme css tags to be added to the overall app}
\usage{
ChangeThemeOutputUI()
}
\value{
theme tags
}
\description{
Theme Switcher - theme css tags to be added to the overall app
}
