% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataAnalysis.R
\name{DataAnalysisIntroUI}
\alias{DataAnalysisIntroUI}
\title{Data Analysis Module - Intro Subtab UI}
\usage{
DataAnalysisIntroUI(strID, strTabName)
}
\arguments{
\item{strID}{namespace ID}

\item{strTabName}{Tab name in sidebar}
}
\value{
tabItem() for run intro subtab of data anlysis module
}
\description{
Data Analysis Module - Intro Subtab UI
}
