% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataAnalysis_runAnalysis.R
\name{DataAnalysisRunAnalysisServer}
\alias{DataAnalysisRunAnalysisServer}
\title{Data Analysis module - Run Analysis server}
\usage{
DataAnalysisRunAnalysisServer()
}
\value{
moduleServer() for run Analysis subtab of data anlysis module
}
\description{
Data Analysis module - Run Analysis server
}
