% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataAnalysis_runAnalysis.R
\name{DataAnalysisRunAnalysisUI}
\alias{DataAnalysisRunAnalysisUI}
\title{Data Analysis module - Run Analysis submodule UI}
\usage{
DataAnalysisRunAnalysisUI(strID, strTabName)
}
\arguments{
\item{strID}{namespace id}

\item{strTabName}{Tab name in sidebar}
}
\value{
tabItem() for run Analysis subtab of data anlysis module
}
\description{
Data Analysis module - Run Analysis submodule UI
}
