% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataAnalysis.R
\name{DataAnalysisSideBarMenu}
\alias{DataAnalysisSideBarMenu}
\title{Data Analysis Shiny Module Sidebar items}
\usage{
DataAnalysisSideBarMenu()
}
\value{
shinydashboard::menuItem() for data analysis module
}
\description{
Data Analysis Shiny Module Sidebar items
}
