% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{SimulationProgramUI}
\alias{SimulationProgramUI}
\title{Simulation shiny module UI - Program subtab}
\usage{
SimulationProgramUI(strID)
}
\arguments{
\item{strID}{namespace id}
}
\value{
UI for Program subtab of simulation module
}
\description{
Simulation shiny module UI - Program subtab
}
