% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{SimulationSideBarMenu}
\alias{SimulationSideBarMenu}
\title{Simulation shiny module sidebar}
\usage{
SimulationSideBarMenu()
}
\value{
shinydashboard::menuItem() for the simulation module
}
\description{
Simulation shiny module sidebar
}
