% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_themeSwitcher.R
\name{ThemeSwitcherServer}
\alias{ThemeSwitcherServer}
\title{Theme Switcher Module Server}
\usage{
ThemeSwitcherServer()
}
\value{
moduleServer() for theme switcher module
}
\description{
Theme Switcher Module Server
}
